Refactoring
Refactored the code when adding decks and
when adding cards.
Now using hashset in AsyncStorage as default
model for saving data and mergeItem based
on the hashset of the card/deck to be
updated/added.
The id is the attribute being used
as the js hashset, and on redux
the changes affected also use
id rather then title to update
the deckList.