Feature
Added firebase support for the app.
Now the user can have its data saved
in firebase backend and know it will
always be saved if he deletes the app.

Added login authentication and now
decks are added depending of the user
that logs in.

separated the firebase functions
in the datahandler folder inside
utils.